% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_rppp.R
\name{grouped_rppp}
\alias{grouped_rppp}
\title{Simulate \link[groupedHyperframe:as.groupedHyperframe]{groupedHyperframe}
with One-and-Only-One \link[spatstat.geom]{ppp}-Hypercolumn}
\usage{
grouped_rppp(..., n, win = square(), envir = parent.frame())
}
\arguments{
\item{...}{see examples in the vignette, for now}

\item{n}{\link[base]{integer} \link[base]{vector},
numbers of \link[spatstat.geom]{ppp.object}s to generate for each set of parameters}

\item{win}{\link[spatstat.geom]{owin.object}}

\item{envir}{\link[base]{environment}}
}
\value{
The function \code{\link[=grouped_rppp]{grouped_rppp()}} returns a \link[groupedHyperframe:as.groupedHyperframe]{groupedHyperframe}.
}
\description{
To simulate a \link[groupedHyperframe:as.groupedHyperframe]{groupedHyperframe}
with one-and-only-one \link[spatstat.geom]{ppp}-hypercolumn.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/nonS3/grouped_rppp_appx.html}
}
\keyword{internal}
