% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_marks.R
\name{r_marks}
\alias{r_marks}
\title{Create Random \code{marks} Generation Function for \link[spatstat.geom]{ppp.object}}
\usage{
r_marks(f)
}
\arguments{
\item{f}{\link[base]{function} of random number generation,
e.g., \link[stats]{rlnorm}, \link[stats]{rnbinom}, etc.
Can also be the function name as a \link[base]{character} scalar.}
}
\value{
Function \code{\link[=r_marks]{r_marks()}} returns a \emph{\strong{\link[base]{function}}},
which generates random marks of a \link[spatstat.geom]{ppp.object}
following the probability distribution specified by argument \code{f}.
The returned \emph{\strong{\link[base]{function}}}
\itemize{
\item {has first parameter \code{x} taking an argument of a \link[spatstat.geom]{ppp.object}.}
\item {returns a \link[spatstat.geom]{ppp.object}.}
}
}
\description{
Create random \code{marks} generation \link[base]{function} for \link[spatstat.geom]{ppp.object}.
}
\examples{
r_marks(rlnorm)
r_marks('rnbinom')

(pp = spatstat.random::rpoispp(lambda = 100))

set.seed(12); pp |>
 r_marks(rlnorm)(sdlog = .5) |>
 r_marks(rnbinom)(size = 5L, prob = .3) |>
 r_marks(rfactor)(prob = c(2,1,3), levels = letters[1:3])
 
}
\keyword{internal}
