% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfactor.R
\name{rfactor}
\alias{rfactor}
\title{Generate Random \link[base]{factor}}
\usage{
rfactor(n, prob, levels = as.character(seq_len(nprob)))
}
\arguments{
\item{n}{\link[base]{integer} scalar}

\item{prob}{\link[base]{numeric} \link[base]{vector}, see function \link[base]{sample.int}}

\item{levels}{\link[base]{character} \link[base]{vector}, see function \link[base]{factor}}
}
\value{
The function \code{\link[=rfactor]{rfactor()}} returns a \link[base]{factor}.
}
\description{
To generate random \link[base]{factor}.
}
\details{
The function \code{\link[=rfactor]{rfactor()}} is a wrapper of \link[base]{sample.int}.
}
\note{
The function \link[stats]{rmultinom} is \strong{not} what we need!
}
\examples{
rfactor(n = 100L, prob = c(4,2,3))
rfactor(n = 100L, prob = c(4,2,3), levels = letters[1:3])
}
\keyword{internal}
