% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{groupedHyperframe.random-package}
\alias{groupedHyperframe.random}
\alias{groupedHyperframe.random-package}
\title{groupedHyperframe.random: Simulated Grouped Hyper Data Frame}
\description{
An intuitive interface to simulate (1) superimposed (marked) point patterns with vectorized parameterization of random point pattern and distribution of marks; and (2) grouped hyper data frame based on population parameters and subject-specific random effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/groupedHyperframe.random}
  \item \url{https://tingtingzhan.quarto.pub/groupedhyperframe/random}
  \item \url{https://tingtingzhan-groupedhyperframe.netlify.app/random}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
