% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{groupedHyperframe.random-package}
\alias{groupedHyperframe.random}
\alias{groupedHyperframe.random-package}
\title{groupedHyperframe.random: Simulated Grouped Hyper Data Frame}
\description{
Simulated Grouped Hyper Data Frame.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID})
}

}
