% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{groupedHyperframe.random-package}
\alias{groupedHyperframe.random}
\alias{groupedHyperframe.random-package}
\title{groupedHyperframe.random: Simulated Grouped Hyper Data Frame}
\description{
An intuitive interface to simulate (1) superimposed (marked) point patterns with vectorized parameterization of random point pattern and distribution of marks; and (2) grouped hyper data frame based on population parameters and subject-specific random effects.
}
\note{
Help files of individual functions are intentionally suppressed in the \code{pdf} manual.
Users are encouraged to get started with

\code{vignette('intro', package = 'groupedHyperframe.random')}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tingtingzhan.quarto.pub/groupedhyperframe/random}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Other contributors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID}) [contributor]
}

}
